Remove manual sliding and improve spacings
